Rename storage to memory
