+++ iwyu.imp
@@ -70,7 +70,7 @@
   { include: ["\"Sts1CobcSw/Edu/Types.hpp\"",                           "public", "<Sts1CobcSw/Edu/Types.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FileSystem/ErrorsAndResult.hpp\"",          "public", "<Sts1CobcSw/FileSystem/ErrorsAndResult.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FileSystem/FileSystem.hpp\"",               "public", "<Sts1CobcSw/FileSystem/FileSystem.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/FileSystem/LfsStorageDevice.hpp\"",         "public", "<Sts1CobcSw/FileSystem/LfsStorageDevice.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/FileSystem/LfsMemoryDevice.hpp\"",          "public", "<Sts1CobcSw/FileSystem/LfsMemoryDevice.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FileSystem/LfsWrapper.hpp\"",               "public", "<Sts1CobcSw/FileSystem/LfsWrapper.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/FramLayout.hpp\"",             "public", "<Sts1CobcSw/FramSections/FramLayout.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/FramSections/RingArray.hpp\"",              "public", "<Sts1CobcSw/FramSections/RingArray.hpp>", "public"] },
